mv tips components to `ui-tips`
