--- packages/app/src/library/Tips/types.ts
@@ -1,9 +1,0 @@
-// Copyright 2026 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
-// SPDX-License-Identifier: GPL-3.0-only
-
-export interface TipProps {
-	title: string
-	description: string[]
-	page?: string
-	onPromptClick: () => void
-}
