+++ packages/ui-tips/package.json
@@ -23,6 +23,9 @@
 		"motion": "12.23.24",
 		"react": "^19.2.7",
 		"react-i18next": "^17.0.8",
+		"ui-buttons": "workspace:*",
+		"ui-core": "workspace:*",
+		"ui-overlay": "workspace:*",
 		"ui-styles": "workspace:*"
 	},
 	"devDependencies": {
